Report host OS when downloading firmware, and allow user to override with --os
I'm not 100% sure if it ever makes a difference, but it seems like
the host OS should be reported accurately when downloading firmware.
However, since its purpose is unclear it may be useful to give the
user an option to override.

Currently, the OS options are 'WINDOWS', 'MAC', or 'LINUX', since those seem
to be the three OS categories which Brother's website reports support for.