Use port 9100 as fallback default for firmware upload
In the case when the printer was *not* autodiscovered (e.g.  IP address
specified explicitly with `-p 192.168.1.234`), the parameter `port` will be
`None`, and the intended default behavior will depend on the local system
database (e.g. `/etc/protocols`) having an entry for `pdl-datastream`.

Add port 9100 as the intended fallback port when this is missing.

See https://github.com/sedrubal/brother_printer_fwupd/commit/831d9858#r141422105